+++ src/hooks/index.ts
@@ -1,2 +1,2 @@
 export { useTitle } from './use-title';
-export { useAppDispatch, useAppSelector } from './hooks';
+export { useAppDispatch, useAppSelector } from './redux-hooks';
